spotify-back/requirements.txt
flask
pymongo
flask-bcrypt
pyjwt
python-dotenv
flask-cors
spotipy
PyJWT